refactor: create new file for getNextPayloadByUserInput and export all from user-input/index file
